Create core game implementation
I created the Die class, capable of rolling itself and storing its result in an encapsulated property.

The Game class holds 3 dice, able to roll them all simultaneously, summate them, output their results and continuously roll dice.

The Program static class holds the main method for the program, instantiating the game class and making it roll dice continuously. The main method also instantiates the testing class for the program, running it after the game is completed. The class also holds an instance of the Random class as a singleton, to prevent over-instantiation errors.

The Testing class holds various tests for the game, ensuring that both the Game class and the Die class work as intended.